Making integration test abstract and reusing it.
