+++ graylog-storage-elasticsearch6/src/test/java/org/graylog/storage/elasticsearch6/SearchesES6IT.java
@@ -1,0 +1,36 @@
+package org.graylog.storage.elasticsearch6;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.graylog2.Configuration;
+import org.graylog2.indexer.results.ScrollResult;
+import org.graylog2.indexer.searches.Searches;
+import org.graylog2.indexer.searches.SearchesIT;
+import org.junit.Before;
+
+import java.util.List;
+
+public class SearchesES6IT extends SearchesIT {
+    @Override
+    public Searches createSearches() {
+        final ScrollResult.Factory scrollResultFactory = new ScrollResult.Factory() {
+            @Override
+            public ScrollResult create(io.searchbox.core.SearchResult initialResult, String query, List<String> fields) {
+                return new ScrollResult(jestClient(), new ObjectMapper(), initialResult, query, fields);
+            }
+
+            @Override
+            public ScrollResult create(io.searchbox.core.SearchResult initialResult, String query, String scroll, List<String> fields) {
+                return new ScrollResult(jestClient(), new ObjectMapper(), initialResult, query, scroll, fields);
+            }
+        };
+
+        return new Searches(
+                indexRangeService,
+                metricRegistry,
+                streamService,
+                indices,
+                indexSetRegistry,
+                new SearchesDriverES6(jestClient(), new Configuration(), scrollResultFactory)
+        );
+    }
+}
