+++ graylog2-server/src/test/java/org/graylog2/indexer/searches/SearchesIT.java
@@ -83,7 +83,7 @@ import static org.mockito.ArgumentMatchers.startsWith;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-public class SearchesIT extends ElasticsearchBaseTest {
+public abstract class SearchesIT extends ElasticsearchBaseTest {
     private static final String REQUEST_TIMER_NAME = "org.graylog2.indexer.searches.Searches.elasticsearch.requests";
     private static final String RANGES_HISTOGRAM_NAME = "org.graylog2.indexer.searches.Searches.elasticsearch.ranges";
 
@@ -145,51 +145,33 @@ public class SearchesIT extends ElasticsearchBaseTest {
     private static final IndexSet indexSet = new TestIndexSet(indexSetConfig);
 
     @Mock
-    private IndexRangeService indexRangeService;
+    protected IndexRangeService indexRangeService;
 
     @Mock
-    private StreamService streamService;
+    protected StreamService streamService;
 
     @Mock
-    private Indices indices;
+    protected Indices indices;
 
     @Mock
-    private IndexSetRegistry indexSetRegistry;
+    protected IndexSetRegistry indexSetRegistry;
 
-    private MetricRegistry metricRegistry;
-    private Searches searches;
+    protected MetricRegistry metricRegistry;
+    protected Searches searches;
 
     @Before
     public void setUp() throws Exception {
         when(indexRangeService.find(any(DateTime.class), any(DateTime.class))).thenReturn(INDEX_RANGES);
         when(indices.getAllMessageFieldsForIndices(any(String[].class))).thenReturn(ImmutableMap.of(INDEX_NAME, Collections.singleton("n")));
         metricRegistry = new MetricRegistry();
-        searches = new Searches(
-                new Configuration(),
-                indexRangeService,
-                metricRegistry,
-                streamService,
-                indices,
-                indexSetRegistry,
-                jestClient(),
-                new ScrollResult.Factory() {
-                    @Override
-                    public ScrollResult create(io.searchbox.core.SearchResult initialResult, String query, List<String> fields) {
-                        return new ScrollResult(jestClient(), new ObjectMapper(), initialResult, query, fields);
-                    }
-
-                    @Override
-                    public ScrollResult create(io.searchbox.core.SearchResult initialResult, String query, String scroll, List<String> fields) {
-                        return new ScrollResult(jestClient(), new ObjectMapper(), initialResult, query, scroll, fields);
-                    }
-                },
-                mock(SearchesDriver.class)
-        );
+        this.searches = createSearches();
     }
+
+    public abstract Searches createSearches();
 
     @Test
     public void testCountWithoutFilter() throws Exception {
-        importFixture("SearchesIT.json");
+        importFixture("org/graylog2/indexer/searches/SearchesIT.json");
 
         CountResult result = searches.count("*", AbsoluteRange.create(
                 new DateTime(2015, 1, 1, 0, 0, DateTimeZone.UTC),
@@ -200,7 +182,7 @@ public class SearchesIT extends ElasticsearchBaseTest {
 
     @Test
     public void testCountWithFilter() throws Exception {
-        importFixture("SearchesIT.json");
+        importFixture("org/graylog2/indexer/searches/SearchesIT.json");
 
         final IndexSetConfig indexSetConfig = IndexSetConfig.builder()
                 .id("id")
@@ -235,7 +217,7 @@ public class SearchesIT extends ElasticsearchBaseTest {
 
     @Test
     public void testCountWithInvalidFilter() throws Exception {
-        importFixture("SearchesIT.json");
+        importFixture("org/graylog2/indexer/searches/SearchesIT.json");
 
         CountResult result = searches.count("*", AbsoluteRange.create(
                 new DateTime(2015, 1, 1, 0, 0, DateTimeZone.UTC),
@@ -247,7 +229,7 @@ public class SearchesIT extends ElasticsearchBaseTest {
 
     @Test
     public void countRecordsMetrics() throws Exception {
-        importFixture("SearchesIT.json");
+        importFixture("org/graylog2/indexer/searches/SearchesIT.json");
 
         CountResult result = searches.count("*", AbsoluteRange.create(
                 new DateTime(2015, 1, 1, 0, 0, DateTimeZone.UTC),
@@ -262,7 +244,7 @@ public class SearchesIT extends ElasticsearchBaseTest {
 
     @Test
     public void testTerms() throws Exception {
-        importFixture("SearchesIT.json");
+        importFixture("org/graylog2/indexer/searches/SearchesIT.json");
 
         TermsResult result = searches.terms("n", 25, "*", AbsoluteRange.create(
                 new DateTime(2015, 1, 1, 0, 0, DateTimeZone.UTC),
@@ -280,7 +262,7 @@ public class SearchesIT extends ElasticsearchBaseTest {
 
     @Test
     public void testTermsWithNonExistingIndex() throws Exception {
-        importFixture("SearchesIT.json");
+        importFixture("org/graylog2/indexer/searches/SearchesIT.json");
 
         final SortedSet<IndexRange> indexRanges = ImmutableSortedSet
                 .orderedBy(IndexRange.COMPARATOR)
@@ -315,7 +297,7 @@ public class SearchesIT extends ElasticsearchBaseTest {
 
     @Test
     public void termsRecordsMetrics() throws Exception {
-        importFixture("SearchesIT.json");
+        importFixture("org/graylog2/indexer/searches/SearchesIT.json");
 
         TermsResult result = searches.terms("n", 25, "*", AbsoluteRange.create(
                 new DateTime(2015, 1, 1, 0, 0, DateTimeZone.UTC),
@@ -334,7 +316,7 @@ public class SearchesIT extends ElasticsearchBaseTest {
 
     @Test
     public void testTermsAscending() throws Exception {
-        importFixture("SearchesIT.json");
+        importFixture("org/graylog2/indexer/searches/SearchesIT.json");
 
         TermsResult result = searches.terms("n", 1, "*", null, AbsoluteRange.create(
             new DateTime(2015, 1, 1, 0, 0, DateTimeZone.UTC),
@@ -349,7 +331,7 @@ public class SearchesIT extends ElasticsearchBaseTest {
 
     @Test
     public void testTermsStats() throws Exception {
-        importFixture("SearchesIT-terms_stats.json");
+        importFixture("org/graylog2/indexer/searches/SearchesIT-terms_stats.json");
 
         TermsStatsResult r = searches.termsStats("f", "n", Searches.TermsStatsOrder.COUNT, 25, "*",
                 AbsoluteRange.create(
@@ -365,7 +347,7 @@ public class SearchesIT extends ElasticsearchBaseTest {
 
     @Test
     public void termsStatsRecordsMetrics() throws Exception {
-        importFixture("SearchesIT-terms_stats.json");
+        importFixture("org/graylog2/indexer/searches/SearchesIT-terms_stats.json");
 
         TermsStatsResult r = searches.termsStats("f", "n", Searches.TermsStatsOrder.COUNT, 25, "*",
                 AbsoluteRange.create(
@@ -386,7 +368,7 @@ public class SearchesIT extends ElasticsearchBaseTest {
 
     @Test
     public void testFieldStats() throws Exception {
-        importFixture("SearchesIT.json");
+        importFixture("org/graylog2/indexer/searches/SearchesIT.json");
 
         FieldStatsResult result = searches.fieldStats("n", "*", AbsoluteRange.create(
                 new DateTime(2015, 1, 1, 0, 0, DateTimeZone.UTC),
@@ -405,7 +387,7 @@ public class SearchesIT extends ElasticsearchBaseTest {
 
     @Test
     public void fieldStatsRecordsMetrics() throws Exception {
-        importFixture("SearchesIT.json");
+        importFixture("org/graylog2/indexer/searches/SearchesIT.json");
 
         FieldStatsResult result = searches.fieldStats("n", "*", AbsoluteRange.create(
                 new DateTime(2015, 1, 1, 0, 0, DateTimeZone.UTC),
@@ -425,7 +407,7 @@ public class SearchesIT extends ElasticsearchBaseTest {
     @Test
     @SuppressWarnings("unchecked")
     public void testHistogram() throws Exception {
-        importFixture("SearchesIT.json");
+        importFixture("org/graylog2/indexer/searches/SearchesIT.json");
 
         final AbsoluteRange range = AbsoluteRange.create(new DateTime(2015, 1, 1, 0, 0, DateTimeZone.UTC).withZone(UTC), new DateTime(2015, 1, 2, 0, 0, DateTimeZone.UTC).withZone(UTC));
         HistogramResult h = searches.histogram("*", Searches.DateHistogramInterval.HOUR, range);
@@ -444,7 +426,7 @@ public class SearchesIT extends ElasticsearchBaseTest {
     @Test
     @SuppressWarnings("unchecked")
     public void testHistogramWithNonExistingIndex() throws Exception {
-        importFixture("SearchesIT.json");
+        importFixture("org/graylog2/indexer/searches/SearchesIT.json");
 
         final SortedSet<IndexRange> indexRanges = ImmutableSortedSet
                 .orderedBy(IndexRange.COMPARATOR)
@@ -479,7 +461,7 @@ public class SearchesIT extends ElasticsearchBaseTest {
 
     @Test
     public void histogramRecordsMetrics() throws Exception {
-        importFixture("SearchesIT.json");
+        importFixture("org/graylog2/indexer/searches/SearchesIT.json");
 
         final AbsoluteRange range = AbsoluteRange.create(new DateTime(2015, 1, 1, 0, 0, DateTimeZone.UTC), new DateTime(2015, 1, 2, 0, 0, DateTimeZone.UTC));
         HistogramResult h = searches.histogram("*", Searches.DateHistogramInterval.MINUTE, range);
@@ -498,7 +480,7 @@ public class SearchesIT extends ElasticsearchBaseTest {
     @Test
     @SuppressWarnings("unchecked")
     public void testFieldHistogram() throws Exception {
-        importFixture("SearchesIT.json");
+        importFixture("org/graylog2/indexer/searches/SearchesIT.json");
 
         final AbsoluteRange range = AbsoluteRange.create(new DateTime(2015, 1, 1, 0, 0, DateTimeZone.UTC).withZone(UTC), new DateTime(2015, 1, 2, 0, 0, DateTimeZone.UTC).withZone(UTC));
         HistogramResult h = searches.fieldHistogram("*", "n", Searches.DateHistogramInterval.HOUR, null, range, false);
@@ -518,7 +500,7 @@ public class SearchesIT extends ElasticsearchBaseTest {
     @Test
     @SuppressWarnings("unchecked")
     public void testFieldHistogramWithMonth() throws Exception {
-        importFixture("SearchesIT.json");
+        importFixture("org/graylog2/indexer/searches/SearchesIT.json");
 
         final AbsoluteRange range = AbsoluteRange.create(new DateTime(2015, 1, 1, 0, 0, DateTimeZone.UTC).withZone(UTC), new DateTime(2015, 1, 2, 0, 0, DateTimeZone.UTC).withZone(UTC));
         HistogramResult h = searches.fieldHistogram("*", "n", Searches.DateHistogramInterval.MONTH, null, range, false);
@@ -536,7 +518,7 @@ public class SearchesIT extends ElasticsearchBaseTest {
     @Test
     @SuppressWarnings("unchecked")
     public void testFieldHistogramWithQuarter() throws Exception {
-        importFixture("SearchesIT.json");
+        importFixture("org/graylog2/indexer/searches/SearchesIT.json");
 
         final AbsoluteRange range = AbsoluteRange.create(new DateTime(2015, 1, 1, 0, 0, DateTimeZone.UTC).withZone(UTC), new DateTime(2015, 1, 2, 0, 0, DateTimeZone.UTC).withZone(UTC));
         HistogramResult h = searches.fieldHistogram("*", "n", Searches.DateHistogramInterval.QUARTER, null, range, false);
@@ -554,7 +536,7 @@ public class SearchesIT extends ElasticsearchBaseTest {
     @Test
     @SuppressWarnings("unchecked")
     public void testFieldHistogramWithYear() throws Exception {
-        importFixture("SearchesIT.json");
+        importFixture("org/graylog2/indexer/searches/SearchesIT.json");
 
         final AbsoluteRange range = AbsoluteRange.create(new DateTime(2015, 1, 1, 0, 0, DateTimeZone.UTC).withZone(UTC), new DateTime(2015, 1, 2, 0, 0, DateTimeZone.UTC).withZone(UTC));
         HistogramResult h = searches.fieldHistogram("*", "n", Searches.DateHistogramInterval.YEAR, null, range, false);
@@ -571,7 +553,7 @@ public class SearchesIT extends ElasticsearchBaseTest {
 
     @Test
     public void fieldHistogramRecordsMetrics() throws Exception {
-        importFixture("SearchesIT.json");
+        importFixture("org/graylog2/indexer/searches/SearchesIT.json");
 
         final AbsoluteRange range = AbsoluteRange.create(new DateTime(2015, 1, 1, 0, 0, DateTimeZone.UTC), new DateTime(2015, 1, 2, 0, 0, DateTimeZone.UTC));
         HistogramResult h = searches.fieldHistogram("*", "n", Searches.DateHistogramInterval.MINUTE, null, range, false);
@@ -760,7 +742,7 @@ public class SearchesIT extends ElasticsearchBaseTest {
 
     @Test
     public void searchDoesNotIncludeJestMetadata() throws Exception {
-        importFixture("SearchesIT.json");
+        importFixture("org/graylog2/indexer/searches/SearchesIT.json");
 
         final AbsoluteRange range = AbsoluteRange.create(new DateTime(2015, 1, 1, 0, 0, DateTimeZone.UTC).withZone(UTC), new DateTime(2015, 1, 2, 0, 0, DateTimeZone.UTC).withZone(UTC));
         final SearchResult searchResult = searches.search("_id:1", range, 0, 0, Sorting.DEFAULT);
@@ -772,7 +754,7 @@ public class SearchesIT extends ElasticsearchBaseTest {
 
     @Test
     public void fieldStatsDoesNotIncludeJestMetadata() throws Exception {
-        importFixture("SearchesIT.json");
+        importFixture("org/graylog2/indexer/searches/SearchesIT.json");
 
         final AbsoluteRange range = AbsoluteRange.create(new DateTime(2015, 1, 1, 0, 0, DateTimeZone.UTC).withZone(UTC), new DateTime(2015, 1, 2, 0, 0, DateTimeZone.UTC).withZone(UTC));
         final FieldStatsResult fieldStatsResult = searches.fieldStats("n", "_id:1", range);
@@ -786,7 +768,7 @@ public class SearchesIT extends ElasticsearchBaseTest {
 
     @Test
     public void searchReturnsCorrectTotalHits() throws Exception {
-        importFixture("SearchesIT.json");
+        importFixture("org/graylog2/indexer/searches/SearchesIT.json");
 
         final AbsoluteRange range = AbsoluteRange.create(new DateTime(2015, 1, 1, 0, 0, DateTimeZone.UTC).withZone(UTC), new DateTime(2015, 1, 2, 0, 0, DateTimeZone.UTC).withZone(UTC));
         final SearchResult searchResult = searches.search("*", range, 5, 0, Sorting.DEFAULT);
@@ -799,7 +781,7 @@ public class SearchesIT extends ElasticsearchBaseTest {
 
     @Test
     public void searchReturnsResultWithSelectiveFields() throws Exception {
-        importFixture("SearchesIT.json");
+        importFixture("org/graylog2/indexer/searches/SearchesIT.json");
 
         final AbsoluteRange range = AbsoluteRange.create(new DateTime(2015, 1, 1, 0, 0, DateTimeZone.UTC).withZone(UTC), new DateTime(2015, 1, 2, 0, 0, DateTimeZone.UTC).withZone(UTC));
         final SearchesConfig searchesConfig = SearchesConfig.builder()
@@ -818,7 +800,7 @@ public class SearchesIT extends ElasticsearchBaseTest {
 
     @Test
     public void scrollReturnsResultWithSelectiveFields() throws Exception {
-        importFixture("SearchesIT.json");
+        importFixture("org/graylog2/indexer/searches/SearchesIT.json");
 
         when(indexSetRegistry.getForIndices(Collections.singleton("graylog_0"))).thenReturn(Collections.singleton(indexSet));
         final AbsoluteRange range = AbsoluteRange.create(new DateTime(2015, 1, 1, 0, 0, DateTimeZone.UTC).withZone(UTC), new DateTime(2015, 1, 2, 0, 0, DateTimeZone.UTC).withZone(UTC));
